#69 one more fix

+++ core/scene_element.cc
@@ -7,6 +7,7 @@
 #include "core/scene_element.h"
 
 #include <cmath>
+#include <cstddef>
 
 #include "core/rotation_matrix.h"
 
@@ -42,7 +43,7 @@ std::vector<Point3D> SceneElement::Vertexes() const {
   std::vector<Point3D> new_vertexes;
   std::vector<Point3D> old_vertexes = mesh_->Vertexes();
 
-  for (int i = 0; i < old_vertexes.size(); i++) {
+  for (size_t i = 0; i < old_vertexes.size(); i++) {
     new_vertexes.push_back(Transform(old_vertexes.at(i)));
   }
